feat(quality): add a flag for very low quality packages
This is used as a custom ranking, which acts much like the `deprecated` flag, in a way that it prefers packages with `false` as the value for `badPackage`. The organisations/owners that I chose to give a lower ranking are:

- [npmtest](https://yarnpkg.com/en/packages?q=npmtest-a&p=1)
- [npmdoc](https://yarnpkg.com/en/packages?q=npmdoc-a&p=1)

These authors are just duplicating packages in a way that provides no value alltogether.