feat(ranking): do tie breaking based on the magnitude of downloads
Doing tie-breaking on the magnitue of the downloads (length of the number) allows for us to put other rules that simply "oh this one is 1 time more downloaded so it should be first"

This also introduced the key `_searchInternal` for data which isn't supposed to be rendered ever. I could put them in `unretrievableAttributes`, but that might be overkill.

We can test this by doing seq=0 and seeing what is happening in the bootstrap index, except if someone else has a better idea of testing it