chore(dependents): remove the data fetching
npm is currently serving us 404 for it (see https://github.com/npm/registry/issues/361)

Removing this is the only solution for now. This will cut down the amount of logs we're doing, and will also make CI green again.

We can revert this commit at a later date when the issue is fixed on npm's side.